# documentation/swagger.yaml
swagger: "2.0"
info: 
  version: "1.0.0"
  title: GameLabz
  description: |
    ## Required software:

    - Java 1.8+
    - PostgreSQL 9+
    - Apache Maven 3+
    
    ## Build:
    CD to the pom folder
    
    ```
    mvn clean package -DskipTests
    ```
    The jar will be in the /target folder.
    
    This SQL script has to be executed once when the server is deployed for the first time:
    
    ```
    drop database if exists "gameLabz";
    create database "gameLabz";
    ```
    
    The required tables will be created by the application.
    
    ## Usage: 
    ```
    java -jar gameLabz-1.0-SNAPSHOT.jar [parameters]
    Parameters:
    --server.port: the server port, default: 8080
    --database.username: the username for the PostgreSQL database, mandatory
    --database.password: the password for the PostgreSQL database, mandatory
    ```
  contact:
    name: Razman Tamas
host: localhost:8080
basePath: /gameLabz
schemes:
  - http
paths:
...
  /admins/{userName}/gameOffers:
    post:
      tags:
        - gameOffers
      summary: An admin creates an offer from existing games
      description: Offers which contain a game or a bundle of games can be created so they can be bought by clients
      operationId: addGameOffer
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: userName
          description: Username of the calling admin user who wants to create a game offer
          required: true
          type: string
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
        - in: body
          name: gameOfferBody
          description: Game offer that needs to be created
          required: true
          schema:
            $ref: "#/definitions/GameOffer"
      responses:
        201:
          description: A new game offer has been successfully created
          schema:
            $ref: "#/definitions/GameOfferWithId"
        400:
          description: Invalid parameters
          headers:
            error_message:
              type: string
              description: The error message
        401:
          description: Admin's session is invalid
          headers:
            error_message:
              type: string
              description: The error message
        403:
          description: Only admins can create game offers
          headers:
            error_message:
              type: string
              description: The error message
  /users/{userName}/gameOffers:
    get:
      tags:
        - gameOffers
      summary: Lists game offers by certain filters 
      operationId: getGameOffersByFilters
      produces:
        - application/json
      parameters:
        - in: path
          name: userName
          description: Username of the user who wants to list the game offers
          required: true
          type: string
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
        - in: query
          name: gameName
          description: To filter game offers by their game's name
          required: false
          type: string
        - in: query
          name: gameGenre
          description: To filter game offers by their game's genre
          required: false
          type: string
        - in: query
          name: offerType
          description: To filter game offers by their type
          required: false
          type: string
        - in: query
          name: isPromotion
          description: To filter game offers which are/aren't in promotion
          required: false
          type: number
          format: double
        - in: query
          name: orderBy
          description: Sorts list after given parameter
          required: false
          type: string
        - in: query
          name: orderDirection
          description: Sets order direction
          required: false
          type: string
          enum:
            - ASC
            - DESC
      responses:
        200:
          description: Listing game offers has been successful
          schema:
            type: array
            items:
              $ref: "#/definitions/GameOfferWithId"
        400:
          description: Invalid parameters
          headers:
            error_message:
              type: string
              description: The error message
        401:
          description: User's session is invalid
          headers:
            error_message:
              type: string
              description: The error message
  /admins/{userName}/gameOffers/{gameOfferId}:
    put:
      tags:
        - gameOffers
      summary: An admin updates an existing game offer
      description: "Finds a game offer by id and updates it. The following fields can be updated: games, price (cash and points), promotion and offer type"
      operationId: updateGameOffer
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: userName
          description: Username of the calling admin user who wants to update the game offer
          required: true
          type: string
        - in: path
          name: gameOfferId
          description: Id of the game offer which needs to be updated
          required: true
          type: integer
          format: int64
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
        - in: body
          name: gameOfferBody
          description: Game offer with new fields to update the old one
          required: true
          schema:
            $ref: "#/definitions/GameOffer"
      responses:
        200:
          description: Game offer has been successfully updated
          schema:
            $ref: "#/definitions/GameOfferWithId"
        400:
          description: Invalid parameters
          headers:
            error_message:
              type: string
              description: The error message
        401:
          description: Admin's session is invalid
          headers:
            error_message:
              type: string
              description: The error message
        403:
          description: Only admins can update game offers
          headers:
            error_message:
              type: string
              description: The error message
          
    delete:
      tags:
        - gameOffers
      summary: An admin removes an existing game offer
      operationId: removeGameOffer
      parameters:
        - in: path
          name: userName
          description: Username of the calling admin user who wants to remove the game offer
          required: true
          type: string
        - in: path
          name: gameOfferId
          description: Id of the game offer which needs to be removed
          required: true
          type: integer
          format: int64
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
      responses:
        200:
          description: Game offer has been successfully removed
        400:
          description: Invalid parameters
          headers:
            error_message:
              type: string
              description: The error message
        401:
          description: Admin's session is invalid
          headers:
            error_message:
              type: string
              description: The error message
        403:
          description: Only admins can remove game offers
          headers:
            error_message:
              type: string
              description: The error message



#----------------------- USER PROFILE ------------------------
...
  /users/{userName}/gameOrder:
    post:
      tags:
        - gameOrders
      summary: A user buys a game or a bundle of games from an offer
      description: An offer has a price and contains a game or a bundle of games which can be purchased by a user with cash, points, keys or it can be sent as a gift to another user
      operationId: makeOrder
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: userName
          description: Username of the user who wants to buy the game(s)
          required: true
          type: string
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
        - in: body
          name: gameOrderBody
          description: The offer that is purchased by the user
          required: true
          schema:
            $ref: "#/definitions/GameOrder"
      responses:
        201:
          description: The offer has been successfully purchased
          schema:
            $ref: "#/definitions/ResponseGameOrder"
        400:
          description: Invalid username
          headers:
            error_message:
              type: string
              description: The error message
        401:
          description: User's session is invalid
          headers:
            error_message:
              type: string
              description: The error message
          
          
          
          
#----------------------- FRIEND REQUEST ------------------------          
...
definitions:
  ResponseUser:
      type: object
      properties:
        userName:
          type: string
        email:
          type: string
        userType:
          type: string
          enum:
            - ADMIN
            - CLIENT
...
  Game:
      type: object
      properties:
        name:
          type: string
        description: 
          type: string
        releaseDate:
          type: integer
          format: int64
        gameGenre:
          type: string
          enum:
            - RPG
            - FPS
            - MOBA
...
  GameOffer:
      type: object
      properties: 
        games:
          type: array
          items:
            $ref: "#/definitions/Game"
        priceCash:
          type: number
          format: double
        pricePoints:
          type: integer
          format: int64
        promotion:
          type: number
          format: double
        offerType:
          type: string
          enum:
            - BUNDLE
            - SINGLE
  GameOfferWithId:
      type: object
      properties:
        id:
          type: integer
          format: int64
        games:
          type: array
          items:
            $ref: "#/definitions/Game"
        priceCash:
          type: number
          format: double
        pricePoints:
          type: integer
          format: int64
        promotion:
          type: number
          format: double
        offerType:
          type: string
          enum:
            - BUNDLE
            - SINGLE
...
  ResponseGameOrder:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: "#/definitions/ResponseUser"
        gameOfferList:
          type: array
          items:
            $ref: "#/definitions/GameOffer"
        ownedGames:
          type: array
          items: 
            $ref: "#/definitions/Game"
        generatedKey:
          $ref: "#/definitions/GeneratedKey"
        dateOfOrder:
          type: integer
          format: int64
  GameOrder:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: "#/definitions/ResponseUser"
        gameOfferList:
          type: array
          items:
            $ref: "#/definitions/GameOffer"
  GeneratedKey:
      type: object
      properties:
        id:
          type: integer
          format: int64
        generatedKey:
          type: string
        user:
          $ref: "#/definitions/ResponseUser"
        game:
          $ref: "#/definitions/Game"
        gameOrder:
          $ref: "#/definitions/GameOrder"
...